Saving current asana integration
