+++ package.json
@@ -3,14 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reach/router": "^1.3.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@types/asana": "^0.14.37",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "antd": "^3.26.8",
+    "asana": "^0.17.3",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
     "react": "^16.12.0",
@@ -42,6 +45,7 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.149",
+    "@types/reach__router": "^1.2.6",
     "@types/react-test-renderer": "^16.9.2",
     "react-test-renderer": "^16.12.0"
   }
